Выполнил задание 8.2
